Refactor external related code for better structure.
Remove pg_exttable.h since the catalog is no longer exist anymore.
Move function declaration in pg_exttable.h into external.h.
Extract related code into external.c which maintain all codes that
can not be moved into external table fdw extension.

Also move the external table orca interface into external.c as work
around. Maybe provide orca fdw routine in future.
