+++ src/backend/access/external/Makefile
@@ -10,7 +10,7 @@ top_builddir = ../../../..
 include $(top_builddir)/src/Makefile.global
 
 OBJS = fileam.o url.o url_curl.o url_file.o url_execute.o url_custom.o \
-	exttable_fdw_shim.o
+	exttable_fdw_shim.o external.o
 
 include $(top_srcdir)/src/backend/common.mk
