+++ src/backend/cdb/cdbpartition.c
@@ -17,6 +17,7 @@
  */
 #include "postgres.h"
 
+#include "access/external.h"
 #include "access/genam.h"
 #include "access/hash.h"
 #include "access/heapam.h"
@@ -29,7 +30,6 @@
 #include "catalog/pg_am.h"
 #include "catalog/pg_constraint.h"
 #include "catalog/pg_constraint_fn.h"
-#include "catalog/pg_exttable.h"
 #include "catalog/pg_inherits.h"
 #include "catalog/pg_inherits_fn.h"
 #include "catalog/pg_type.h"
