+++ src/backend/gpopt/translate/CTranslatorDXLToPlStmt.cpp
@@ -21,7 +21,6 @@
 #include "nodes/primnodes.h"
 #include "catalog/gp_policy.h"
 #include "catalog/pg_collation.h"
-#include "catalog/pg_exttable.h"
 #include "cdb/cdbutil.h"
 #include "cdb/cdbvars.h"
 #include "cdb/partitionselection.h"
