+++ src/include/access/exttable_fdw_shim.h
@@ -24,6 +24,4 @@ extern Datum pg_exttable(PG_FUNCTION_ARGS);
 
 extern void gp_exttable_permission_check(PG_FUNCTION_ARGS);
 
-extern ForeignScan *create_foreignscan_for_external_table(Oid relid, Index scanrelid, List *qual, List *targetlist);
-
 #endif   /* EXTTABLE_FDW_SHIM_H */
